# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FuelGo Islamabad - Fuel Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers/profile:
    patch:
      operationId: updateCustomerProfile
      tags: [users]
      summary: Update customer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/favorites:
    get:
      operationId: getCustomerFavorites
      tags: [users]
      summary: Get customer favorite pumps
      responses:
        "200":
          description: Favorite pumps list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetrolPump"
    post:
      operationId: addFavoritePump
      tags: [users]
      summary: Add a pump to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /customers/favorites/{pumpId}:
    delete:
      operationId: removeFavoritePump
      tags: [users]
      summary: Remove a pump from favorites
      parameters:
        - name: pumpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed from favorites

  # ─── PUMPS ──────────────────────────────────────────────────────────────────
...
  /pumps/{pumpId}/fuel:
    get:
      operationId: getPumpFuelPrices
      tags: [fuel]
      summary: Get fuel types and prices for a pump
      parameters:
        - name: pumpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fuel prices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuelPrice"

    put:
      operationId: updatePumpFuelPrices
      tags: [fuel]
      summary: Update fuel prices for own pump (pump owner only)
      parameters:
        - name: pumpId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FuelPricesUpdate"
      responses:
        "200":
          description: Fuel prices updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuelPrice"

  # ─── ORDERS ─────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (customer sees own orders; pump owner sees orders for their pump)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new fuel delivery order (customer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (pump owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/cancel:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order (customer can cancel pending orders)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Cannot cancel at this stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── REVIEWS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, fullName, phone, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
    CustomerProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        profilePicture:
          type: string
    FavoriteInput:
      type: object
      required: [pumpId]
      properties:
        pumpId:
          type: integer

    # ── Pump ────────────────────────────────────────────────────────────────
    PetrolPump:
      type: object
      required: [id, businessName, ownerName, phone, address, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        businessName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        operatingHoursStart:
          type: ["string", "null"]
        operatingHoursEnd:
          type: ["string", "null"]
        deliveryRadius:
          type: ["number", "null"]
        deliveryCharges:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        rejectionReason:
          type: ["string", "null"]
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: ["integer", "null"]
        isOpen:
          type: ["boolean", "null"]
        fuelPrices:
          type: array
          items:
            $ref: "#/components/schemas/FuelPrice"
        createdAt:
          type: string
...
    FuelPrice:
      type: object
      required: [id, pumpId, fuelType, pricePerLiter, isAvailable]
      properties:
        id:
          type: integer
        pumpId:
          type: integer
        fuelType:
          type: string
          enum: [petrol, diesel, hi_octane]
        pricePerLiter:
          type: number
        isAvailable:
          type: boolean
        updatedAt:
          type: string
    FuelPriceEntry:
      type: object
      required: [fuelType, pricePerLiter, isAvailable]
      properties:
        fuelType:
          type: string
          enum: [petrol, diesel, hi_octane]
        pricePerLiter:
          type: number
        isAvailable:
          type: boolean
    FuelPricesUpdate:
      type: object
      required: [prices]
      properties:
        prices:
          type: array
          items:
            $ref: "#/components/schemas/FuelPriceEntry"

    # ── Orders ──────────────────────────────────────────────────────────────
    OrderInput:
      type: object
      required: [pumpId, fuelType, quantityLiters, deliveryAddress]
      properties:
        pumpId:
          type: integer
        fuelType:
          type: string
          enum: [petrol, diesel, hi_octane]
        quantityLiters:
          type: number
        deliveryAddress:
          type: string
        deliveryLat:
          type: ["number", "null"]
        deliveryLng:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, preparing, out_for_delivery, delivered, cancelled]
        notes:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, customerId, pumpId, fuelType, quantityLiters, pricePerLiter, deliveryCharges, totalAmount, status, deliveryAddress, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        pumpId:
          type: integer
        fuelType:
          type: string
          enum: [petrol, diesel, hi_octane]
        quantityLiters:
          type: number
        pricePerLiter:
          type: number
        deliveryCharges:
          type: number
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, accepted, rejected, preparing, out_for_delivery, delivered, cancelled]
        deliveryAddress:
          type: string
        deliveryLat:
          type: ["number", "null"]
        deliveryLng:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        statusNotes:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        pumpName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    # ── Reviews ─────────────────────────────────────────────────────────────
...