# lib/api-zod/src/generated/types/fuelPriceEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FuelGo Islamabad - Fuel Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FuelPriceEntryFuelType } from './fuelPriceEntryFuelType';

export interface FuelPriceEntry {
  fuelType: FuelPriceEntryFuelType;
  pricePerLiter: number;
  isAvailable: boolean;
}